@@ -20,70 +20,72 @@
     "test": "make test"
   },
   "dependencies": {
-    "argparse": "^1.0.10",
-    "autoprefixer": "^6.7.7",
-    "babel-preset-es2015": "^6.24.1",
+    "@babel/core": "^7.2.2",
+    "@babel/preset-env": "^7.3.1",
+    "argparse": "^1.0.9",
+    "autoprefixer": "^9.4.7",
     "babelfish": "^1.1.2",
-    "babelify": "^7.2.0",
-    "bagjs": "^2.0.2",
-    "base64-js": "~1.1.1",
+    "babelify": "^10.0.0",
+    "bagjs": "^3.0.0",
+    "base64-js": "^1.2.0",
     "bluebird": "^3.5.1",
-    "bluebird-co": "^2.2.0",
     "bootstrap": "3.2.0",
-    "bootstrap-material-design": "^4.1.1",
-    "browser-pack": "^6.0.4",
-    "browserify": "^13.3.0",
+    "browser-pack": "^6.0.2",
+    "browserify": "^16.1.0",
     "cacheify": "^0.4.2",
-    "compression": "^1.7.2",
-    "cookie": "^0.3.1",
-    "csswring": "^5.1.1",
-    "es6-promise": "^3.3.1",
+    "clean-css": "^4.2.1",
+    "compression": "^1.6.1",
+    "cookie": "^0.4.0",
+    "email-regex": "^4.0.0",
     "event-wire": "^3.4.0",
-    "formidable": "~ 1.0.15",
+    "fontello-cli": "^0.5.0",
     "glob": "^7.1.2",
-    "is-my-json-valid": "^2.17.2",
-    "jade": "~ 1.11.0",
-    "javascript-state-machine": "~ 2.3.5",
-    "jquery": "^3.3.1",
-    "js-yaml": "^3.11.0",
-    "jszip": "^3.1.5",
-    "knockout": "^3.4.2",
-    "less": "^2.7.3",
-    "level": "^1.7.0",
-    "level-exists": "^2.0.1",
+    "is-my-json-valid": "^2.16.1",
+    "javascript-state-machine": "^2.4.0",
+    "jquery": "^3.2.1",
+    "js-yaml": "^3.9.0",
+    "jstransformer-stylus": "^1.3.0",
+    "jszip": "^3.1.3",
+    "knockout": "~ 3.4.2",
+    "less": "^3.9.0",
+    "level": "^6.0.0",
+    "level-exists": "^2.0.2",
     "level-ttl": "^3.1.1",
-    "lodash": "^4.17.5",
-    "log4js": "~ 0.6.31",
+    "lie": "^3.1.1",
+    "lodash": "^4.17.4",
+    "log4js": "^6.1.2",
     "mdurl": "^1.0.1",
-    "mkdirp": "^0.5.1",
-    "mz": "^2.7.0",
-    "node-sass": "^3.13.1",
-    "npm": "^3.10.10",
-    "pointer": "^1.0.2",
-    "postcss": "^5.2.18",
+    "mkdirp": "^1.0.3",
+    "multiparty": "^4.1.3",
+    "mz": "^2.6.0",
+    "negotiator": "^0.6.1",
+    "pointer": "^1.0.1",
+    "postcss": "^7.0.14",
+    "pug": "^2.0.0-rc.2",
+    "pug-runtime": "^2.0.3",
     "puncher": "^1.1.1",
-    "readable-stream": "^2.3.5",
-    "resolve": "~ 1.1.7",
-    "rimraf": "^2.6.2",
-    "send": "^0.14.2",
-    "stylus": "~ 0.54.5",
-    "svg2ttf": "^4.1.0",
+    "raw-body": "^2.2.0",
+    "readable-stream": "^3.1.1",
+    "resolve": "^1.5.0",
+    "send": "^0.17.1",
+    "serialize-error": "^5.0.0",
+    "stylus": "^0.54.5",
+    "svg2ttf": "^4.0.3",
     "svgpath": "^2.2.1",
     "ttf2eot": "^2.0.0",
     "ttf2woff": "^2.0.1",
-    "ttf2woff2": "^2.0.3",
     "typeahead.js": "0.11.1",
-    "uglify-js": "^2.8.29",
-    "uglifyify": "^3.0.4",
-    "write-file-atomic": "^1.3.4",
-    "xmldom": "^0.1.27"
+    "uglify-js": "^3.1.10",
+    "uglifyify": "^5.0.0",
+    "wawoff2": "^1.0.1",
+    "write-file-atomic": "^3.0.0",
+    "xmldom": "~0.2.1"
   },
   "devDependencies": {
-    "acorn": "^6.0.2",
-    "eslint": "^3.19.0",
-    "eslint-plugin-nodeca": "~1.0.3",
-    "mocha": "^3.5.3",
-    "supertest": "^2.0.1"
+    "acorn": "^7.1.0",
+    "eslint": "^6.8.0",
+    "mocha": "^7.0.1",
+    "supertest": "^4.0.2"
   },
   "browserify": {
     "transform": [
@@ -91,10 +93,15 @@
         "babelify",
         {
           "presets": [
-            "es2015"
+            "@babel/preset-env"
           ]
         }
       ]
     ]
-  }
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "Firefox ESR"
+  ]
 }
