@@ -1,10 +1,9 @@
 'use strict';
 
 
-const _         = require('lodash');
-const path      = require('path');
-const co        = require('bluebird-co').co;
-const Promise   = require('bluebird');
+const _       = require('lodash');
+const Promise = require('bluebird');
+const path    = require('path');
 
 
 function Asset(bundler, options) {
@@ -40,7 +39,9 @@ Object.defineProperty(Asset.prototype, 'digestPath', {
 
     const path_obj = path.parse(this.destPath);
 
-    return path.join(path_obj.dir, path_obj.name + '-' + this.digest + path_obj.ext);
+    // convert backward slashes on windows to forward slashes to simplify further code
+    return path_obj.dir.replace(new RegExp('\\' + path.sep, 'g'), '/') + '/' +
+           path_obj.name + '-' + this.digest + path_obj.ext;
   }
 });
 
@@ -54,7 +55,7 @@ Object.defineProperty(Asset.prototype, 'buffer', {
     }
 
     if (this.__source__ !== null) {
-      this.__buffer__ = new Buffer(this.source);
+      this.__buffer__ = Buffer.from(this.source);
       return this.__buffer__;
     }
 
@@ -97,22 +98,20 @@ Object.defineProperty(Asset.prototype, 'length', {
 });
 
 
-Asset.prototype.__run_plugins__ = co.wrap(function* () {
+Asset.prototype.__run_plugins__ = async function () {
   let context = {
     bundler: this.__bundler__,
     asset: this
   };
 
-  for (let i = 0; i < this.plugins.length; i++) {
-    let plugin = this.plugins[i];
-
+  for (let plugin of this.plugins) {
     if (_.isString(plugin)) {
-      yield this.__bundler__.__plugins__[plugin](context);
+      await this.__bundler__.__plugins__[plugin](context);
     } else {
-      yield plugin(context);
+      await plugin(context);
     }
   }
-});
+};
 
 
 Asset.prototype.__throw_if_not_resolved__ = function () {
@@ -198,7 +197,7 @@ Asset.prototype.__restore_cache__ = function () {
           .then(() => {
             if (this.dependenciesDigest !== cache.dependenciesDigest) return false;
 
-            if (cache.buffer) this.__buffer__ = new Buffer(cache.buffer);
+            if (cache.buffer) this.__buffer__ = Buffer.from(cache.buffer);
             else this.__source__ = cache.source;
 
             this.digest = cache.dependenciesDigest;
