@@ -1,15 +1,14 @@
 'use strict';
 
 
-const zlib    = require('mz/zlib');
-const mkdirp  = require('mkdirp');
-const fs      = require('fs');
-const crypto  = require('crypto');
-const _       = require('lodash');
-const path    = require('path');
-const co      = require('bluebird-co').co;
-const Promise = require('bluebird');
-const write   = Promise.promisify(require('write-file-atomic'));
+const mkdirp        = require('mkdirp');
+const fs            = require('fs');
+const crypto        = require('crypto');
+const _             = require('lodash');
+const path          = require('path');
+const { promisify } = require('util');
+const gzip          = promisify(require('zlib').gzip);
+const write         = promisify(require('write-file-atomic'));
 
 
 const asset_classes = [
@@ -28,10 +27,11 @@ const plugins = [
   'sass',
   'auto',
   'jade',
+  'pug',
   'macros',
   'autoprefixer',
   'uglifyjs',
-  'csswring'
+  'clean-css'
 ].reduce((acc, name) => {
   acc[name] = require(path.join(__dirname, 'plugins', name));
   return acc;
@@ -154,12 +154,12 @@ Bundler.prototype.readFile = function (filename, encoding) {
 
 // Compile assets and create manifest file
 //
-Bundler.prototype.compile = co.wrap(function* () {
+Bundler.prototype.compile = async function () {
   // Resolve public assets only. Dependencies will ve resolved automatically.
   const public_assets = _.values(this.__assets__).filter(asset => !asset.virtual);
 
   // Resolve public assets
-  yield public_assets.map(asset => asset.resolve());
+  await Promise.all(public_assets.map(asset => asset.resolve()));
 
   // Write assets
   for (let i = 0; i < public_assets.length; i++) {
@@ -176,14 +176,14 @@ Bundler.prototype.compile = co.wrap(function* () {
     let buffer = asset.buffer;
 
     if (this.sourceMaps && asset.sourceMap()) {
-      buffer = Buffer.concat([ buffer, new Buffer(asset.mappingUrlComment()) ]);
+      buffer = Buffer.concat([ buffer, Buffer.from(asset.mappingUrlComment()) ]);
     }
 
     mkdirp.sync(path.dirname(target));
 
     // Can be done in parallel
 
-    yield Promise.all([
+    await Promise.all([
       // Write asset
       write(target, buffer),
 
@@ -191,7 +191,7 @@ Bundler.prototype.compile = co.wrap(function* () {
       Promise.resolve().then(() => {
         if (!this.compress) return;
 
-        return zlib.gzip(buffer).then(res => write(target + '.gz', res));
+        return gzip(buffer).then(res => write(target + '.gz', res));
       }),
 
       // Write source map
@@ -205,7 +205,7 @@ Bundler.prototype.compile = co.wrap(function* () {
       Promise.resolve().then(() => {
         if (!asset.sourceMap() || !this.__options__.compress) return;
 
-        return zlib.gzip(')]}\'\n' + asset.sourceMap())
+        return gzip(')]}\'\n' + asset.sourceMap())
           .then(res => write(target + '.map.gz', res));
       })
     ]);
@@ -224,7 +224,7 @@ Bundler.prototype.compile = co.wrap(function* () {
                   .valueOf();
 
   return manifest;
-});
+};
 
 
 Bundler.DependencyError = Bundler.prototype.DependencyError = require('./utils/dependency_error');
