@@ -1,11 +1,9 @@
 // Show wire listeners
 //
-
 'use strict';
 
 
-const _  = require('lodash');
-const co = require('bluebird-co').co;
+const _       = require('lodash');
 
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -41,48 +39,46 @@ module.exports.commandLineArguments = [
 ];
 
 
-module.exports.run = function (N, args) {
+module.exports.run = async function (N, args) {
 
-  return co(function* () {
-    // Reduce log level
-    N.logger.setLevel('info');
+  // Reduce log level
+  N.logger.setLevel('info');
 
-    yield Promise.resolve()
-      .then(() => N.wire.emit('init:models', N))
-      .then(() => N.wire.emit('init:bundle', N))
-      .then(() => N.wire.emit('init:services', N));
+  await Promise.resolve()
+    .then(() => N.wire.emit('init:models', N))
+    .then(() => N.wire.emit('init:bundle', N))
+    .then(() => N.wire.emit('init:services', N));
 
-    /*eslint-disable no-console*/
-    console.log('\n');
+  /*eslint-disable no-console*/
+  console.log('\n');
 
-    _.forEach(N.wire.stat(), function (hook) {
-      // try to filter by pattern, if set
-      if (args.mask && (hook.name.indexOf(args.mask) === -1)) {
-        return;
-      }
+  _.forEach(N.wire.stat(), function (hook) {
+    // try to filter by pattern, if set
+    if (args.mask && (hook.name.indexOf(args.mask) === -1)) {
+      return;
+    }
 
-      if (args.short) {
-        // short formst
-        console.log(`- ${hook.name}`);
-      } else {
-        // long format
-        console.log(`\n${hook.name} -->\n`);
-        _.forEach(hook.listeners, function (h) {
-          let flags = [];
+    if (args.short) {
+      // short formst
+      console.log(`- ${hook.name}`);
+    } else {
+      // long format
+      console.log(`\n${hook.name} -->\n`);
+      _.forEach(hook.listeners, function (h) {
+        let flags = [];
 
-          if (h.ensure)   flags.push('permanent');
-          if (h.parallel) flags.push('parallel');
+        if (h.ensure)   flags.push('permanent');
+        if (h.parallel) flags.push('parallel');
 
-          console.log(
-            `  - [${h.priority}] ${h.name}     (cnt: ${h.ncalled})` +
-            (flags.length ? '    !' + flags.join(', ') : '')
-          );
-        });
-      }
-    });
-
-    console.log('\n');
-
-    return N.wire.emit('exit.shutdown');
+        console.log(
+          `  - [${h.priority}] ${h.name}     (cnt: ${h.ncalled})` +
+          (flags.length ? '    !' + flags.join(', ') : '')
+        );
+      });
+    }
   });
+
+  console.log('\n');
+
+  await N.wire.emit('exit.shutdown');
 };
