@@ -4,11 +4,10 @@
 'use strict';
 
 
-const path         = require('path');
-const _            = require('lodash');
-const co           = require('bluebird-co').co;
-const Mocha        = require('mocha');
-const glob         = require('glob');
+const _       = require('lodash');
+const glob    = require('glob');
+const Mocha   = require('mocha');
+const path    = require('path');
 
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -46,62 +45,76 @@ module.exports.commandLineArguments = [
 
 ////////////////////////////////////////////////////////////////////////////////
 
-module.exports.run = function (N, args) {
+module.exports.run = async function (N, args) {
 
-  return co(function* () {
-    if (!process.env.NODECA_ENV) {
-      throw 'You must provide NODECA_ENV in order to run nodeca test';
+  if (!process.env.NODECA_ENV) {
+    throw 'You must provide NODECA_ENV in order to run nodeca test';
+  }
+
+  // Expose N to globals for tests
+  global.TEST = { N };
+
+  await Promise.resolve()
+    .then(() => N.wire.emit('init:models', N))
+    .then(() => N.wire.emit('init:bundle', N))
+    .then(() => N.wire.emit('init:services', N))
+    .then(() => N.wire.emit('init:tests', N));
+
+  let mocha        = new Mocha({ timeout: 60000 });
+  let applications = N.apps;
+
+  mocha.reporter('spec');
+  // mocha.ui('bdd');
+
+  // if app set, check that it's valid
+  if (args.app) {
+    if (!_.find(applications, app => app.name === args.app)) {
+      let msg = `Invalid application name: ${args.app}\n` +
+          'Valid apps are:  ' + _.map(applications, app => app.name).join(', ');
+
+      throw msg;
     }
+  }
 
-    // Expose N to globals for tests
-    global.TEST = { N };
+  let testedApps = [];
 
-    yield Promise.resolve()
-      .then(() => N.wire.emit('init:models', N))
-      .then(() => N.wire.emit('init:bundle', N))
-      .then(() => N.wire.emit('init:services', N))
-      .then(() => N.wire.emit('init:tests', N));
+  if (args.app) {
+    // if app is defined, add it and all modules recursively
+    applications.filter(app => args.app === app.name).forEach(function addApp(app) {
+      testedApps.push(app);
 
-    let mocha        = new Mocha({ timeout: 60000 });
-    let applications = N.apps;
+      app.config.modules.forEach(m => {
+        let modulePath = path.join(app.root, m);
 
-    mocha.reporter('spec');
-    // mocha.ui('bdd');
-
-    // if app set, check that it's valid
-    if (args.app) {
-      if (!_.find(applications, app => app.name === args.app)) {
-        let msg = `Invalid application name: ${args.app}` +
-            'Valid apps are:  ' + _.map(applications, app => app.name).join(', ');
-
-        throw msg;
-      }
-    }
-
-    _.forEach(applications, app => {
-      if (!args.app || args.app === app.name) {
-        glob.sync('**', { cwd: app.root + '/test' })
-          // skip files when
-          // - filename starts with _, e.g.: /foo/bar/_baz.js
-          // - dirname in path starts _, e.g. /foo/_bar/baz.js
-          .filter(name => !/^[._]|\\[._]|\/[_.]/.test(name))
-          .forEach(file => {
-            // try to filter by pattern, if set
-            if (args.mask && path.basename(file).indexOf(args.mask) === -1) {
-              return;
-            }
-
-            if ((/\.js$/).test(file) && path.basename(file)[0] !== '.') {
-              mocha.files.push(`${app.root}/test/${file}`);
-            }
-          });
-      }
+        applications.filter(app => app.root === modulePath).forEach(addApp);
+      });
     });
+  } else {
+    // otherwise test all applications
+    testedApps = applications;
+  }
 
-    yield new Promise((resolve, reject) => {
-      mocha.run(err => (err ? reject(err) : resolve()));
-    });
+  _.forEach(testedApps, app => {
+    glob.sync('**', { cwd: app.root + '/test' })
+      // skip files when
+      // - filename starts with _, e.g.: /foo/bar/_baz.js
+      // - dirname in path starts _, e.g. /foo/_bar/baz.js
+      .filter(name => !/^[._]|\\[._]|\/[_.]/.test(name))
+      .forEach(file => {
+        // try to filter by pattern, if set
+        if (args.mask && path.basename(file).indexOf(args.mask) === -1) {
+          return;
+        }
 
-    return N.wire.emit('exit.shutdown');
+        if ((/\.js$/).test(file) && path.basename(file)[0] !== '.') {
+          mocha.files.push(`${app.root}/test/${file}`);
+        }
+      });
   });
+
+  await new Promise((resolve, reject) => {
+    mocha.run(err => (err ? reject(err) : resolve()));
+  });
+
+  await N.wire.emit('exit.shutdown');
 };
