@@ -1,17 +1,10 @@
-// Start worker (queue part)
+// Start HTTP worker
 //
 
 'use strict';
 
 
 ////////////////////////////////////////////////////////////////////////////////
-
-
-module.exports.parserParameters = {
-  addHelp: true,
-  help: 'start worker (queue)',
-  description: 'Start worker (queue)'
-};
 
 
 module.exports.commandLineArguments = [];
@@ -23,5 +16,5 @@ module.exports.run = function (N/*, args*/) {
   return Promise.resolve()
     .then(() => N.wire.emit('init:models', N))
     .then(() => N.wire.emit('init:bundle', N))
-    .then(() => N.wire.emit('init:server.worker-queue', N));
+    .then(() => N.wire.emit('init:services.www', N));
 };
