@@ -14,14 +14,12 @@
 // data - event payload
 //
 function handleAction(apiPath, data) {
-  N.loader.loadAssets(apiPath.split('.')[0], function () {
-    if (N.wire.has(apiPath)) {
-      N.wire.emit(apiPath, data)
-        .catch(err => N.wire.emit('error', err));
-    } else {
-      N.logger.error('Unknown client Wire channel: %s', apiPath);
-    }
-  });
+  if (N.wire.has(apiPath)) {
+    N.wire.emit(apiPath, data)
+      .catch(err => N.wire.emit('error', err));
+  } else {
+    N.logger.error('Unknown client Wire channel: %s', apiPath);
+  }
 }
 
 
