@@ -13,7 +13,6 @@ const http        = require('http');
 const https       = require('https');
 const os          = require('os');
 const path        = require('path');
-const querystring = require('querystring');
 const url         = require('url');
 const env         = require('../../../system/env');
 const WorkerPool  = require('../../../system/worker_pool');
@@ -135,28 +134,19 @@ function getRequestRemoteAddress(req, isForwarded) {
 }
 
 
-// Parses URL query string and resolves numbers and booleans.
-//
-function parseQueryString(inputQuery) {
-  if (!inputQuery) return {};
-
-  return querystring.parse(String(inputQuery));
-}
-
-
 module.exports = function (N) {
   ////////////////////////////////////////////////////////////////////////////////
   // Start http worker pool
   //
   // In cluster mode it forks N http workers, in single mode it emits
-  // `init:server.worker-http` instead
+  // `init:server.www` instead
   //
   let active_pools = [];
 
-  N.wire.on('init:services', { parallel: true }, function* worker_pool_http_init(N) {
-    let fork = N.config.fork;
+  N.wire.on('init:services', { parallel: true }, async function worker_pool_http_init(N) {
+    let fork = (N.config.fork || {}).www;
 
-    if (fork === 'auto' || fork === true || _.isUndefined(fork)) {
+    if (fork === 'auto') {
       fork = os.cpus().length;
     } else {
       fork = +fork || 0;
@@ -164,11 +154,11 @@ module.exports = function (N) {
 
     if (!cluster.isMaster || fork < 1) {
       // single process mode
-      yield N.wire.emit('init:services.worker-http', N);
+      await N.wire.emit('init:services.www', N);
       return;
     }
 
-    let pool = new WorkerPool('worker-http', fork);
+    let pool = new WorkerPool('www', fork);
 
     active_pools.push(pool);
 
@@ -177,21 +167,23 @@ module.exports = function (N) {
     pool.on('worker:online', pid         => N.logger.info(`Worker ${pid} is running`));
     pool.on('worker:exit',   (pid, code) => N.logger.info(`Worker ${pid} exited with status ${code}`));
 
-    N.wire.on('exit.shutdown', { ensure: true, parallel: true }, function shutdown_http_pool(__, callback) {
-      pool.once('exit', callback);
+    N.wire.on('exit.shutdown', { ensure: true, parallel: true }, function shutdown_http_pool() {
       pool.shutdown();
+
+      return new Promise(resolve => pool.once('exit', resolve));
     });
 
-    N.wire.on('exit.terminate', { ensure: true, parallel: true }, function terminate_http_pool(__, callback) {
-      pool.once('exit', callback);
+    N.wire.on('exit.terminate', { ensure: true, parallel: true }, function terminate_http_pool() {
       pool.terminate();
+
+      return new Promise(resolve => pool.once('exit', resolve));
     });
 
     N.wire.on('reload', function reload_http_pool() {
       pool.reload();
     });
 
-    yield new Promise((resolve, reject) => {
+    await new Promise((resolve, reject) => {
       // Wait for either 'online' or 'error' events,
       // whichever comes first
       //
@@ -286,8 +278,7 @@ module.exports = function (N) {
       }
 
       parsedUrl = url.parse(req.url);
-      fullUrl   = protocol + '://' + req.headers.host + parsedUrl.pathname;
-      req.query = parseQueryString(parsedUrl.query);
+      fullUrl   = protocol + '://' + req.headers.host + parsedUrl.path;
 
       //
       // Choose a responder to use.
@@ -329,7 +320,7 @@ module.exports = function (N) {
   }
 
 
-  N.wire.on('init:services.worker-http', function* server_bind(N) {
+  N.wire.on('init:services.www', async function server_bind(N) {
     let bindings = extractBindings(N.config.bind, N.mainApp, N.logger);
     let bindingNames = _.keys(bindings);
 
@@ -405,7 +396,7 @@ module.exports = function (N) {
 
     // Bind web servers
     for (let i = 0; i < bindingNames.length; i++) {
-      yield bindServer(bindingNames[i]);
+      await bindServer(bindingNames[i]);
     }
   });
 
