@@ -5,19 +5,22 @@
 'use strict';
 
 
-var http = require('http');
-var _    = require('lodash');
+const http = require('http');
+const _    = require('lodash');
 
 
 ////////////////////////////////////////////////////////////////////////////////
 
+// Note, we also listen `responder:bin` to unify logging & error responding
+
+const CHANNELS = [ 'responder:http', 'responder:rpc', 'responder:bin' ];
 
 module.exports = function (N) {
 
-  N.wire.after([ 'responder:http', 'responder:rpc' ], { priority: 50 }, function error_process(env) {
-    var err = env.err;
+  N.wire.after(CHANNELS, { priority: 50 }, function error_process(env) {
+    if (!env.err) return;
 
-    if (!err) return;
+    let err = env.err_orig = env.err;
 
     // Extend sugared errors
     // Example: next(404);
@@ -50,7 +53,7 @@ module.exports = function (N) {
     // Still no code -> we got Error object or string
     // Example: next(new Error('Fatal fuckup'))
 
-    var e = {
+    let e = {
       code: N.io.APP_ERROR
     };
 
