@@ -1,60 +1,36 @@
 // Prepare http request for server chain
 // - find method (with router)
 // - parse parameters
-
-
+// - parse multipart post data
+//
+// WARNING!!! Here are NO built-in CSRF protection like in RPC.
+// If you plan to enable POST requests in router, be sure to
+// care about security.
+//
 'use strict';
 
 
-var _           = require('lodash');
-var http        = require('http');
+const _          = require('lodash');
+const Promise    = require('bluebird');
+const multiparty = require('multiparty');
+
+
+const MAX_FIELDS_SIZE = 100 * 1024;       // 100kb
+const MAX_FILES_SIZE  = 10 * 1024 * 1024; // 10mb
 
 
 module.exports = function (N) {
-  //
-  // Request logger for http requests
-  //
-  function log(env) {
-    var err = env.err,
-        req = env.origin.req,
-        message = http.STATUS_CODES[env.status],
-        logger = N.logger.getLogger('http@' + env.method),
-        level;
-
-    if (N.io.APP_ERROR <= env.status) {
-      message = err.stack || err.message || JSON.stringify(err);
-      level   = 'fatal';
-    } else if (N.io.BAD_REQUEST <= env.status && N.io.APP_ERROR > env.status) {
-      message = err.message || http.STATUS_CODES[err.code];
-      level = 'error';
-    } else {
-      level = 'info';
-    }
-
-    logger[level]('%s - "%s %s HTTP/%s" %d "%s" - %s',
-                  env.req.ip,
-                  req.method,
-                  req.url,
-                  req.httpVersion,
-                  env.status,
-                  req.headers['user-agent'] || '',
-                  message);
-  }
 
   //
-  // Init envirement for http
+  // Init environment for http
   //
-  N.wire.before('responder:http', function http_prepare(env) {
+
+  N.wire.before('responder:http', async function http_prepare(env) {
     var req        = env.origin.req,
         httpMethod = req.method.toLowerCase(),
-        match      = env.req.matched, // N.router.match(req.fullUrl)
-        // mix GET QUERY params (part after ? in URL) and params from router
-        // params from router take precedence
-        params     = _.assign({}, req.query, (match || {}).params);
+        match      = env.req.matched; // N.router.match(req.fullUrl)
 
-    env.log_request = log;
-
-    env.params = params;
+    env.params = (match || {}).params;
 
     // Nothing matched -> error
     if (!match) {
@@ -69,5 +45,46 @@ module.exports = function (N) {
     }
 
     env.method = match.meta.methods[httpMethod];
+
+    if (req.method !== 'POST') return;
+
+    // Parse body for POST request
+
+    //
+    // Check request size early by header and terminate immediately for big data
+    //
+    let length = parseInt((req.headers['content-length'] || '0'), 10);
+
+    if (!length || isNaN(length)) {
+      env.err = N.io.LENGTH_REQUIRED;
+      return;
+    }
+
+    let err = null;
+
+    // Now we understand only multipart form requests
+    if (/^multipart\/form-data(?:;|$)/i.test(req.headers['content-type'])) {
+
+      let form = new multiparty.Form({
+        maxFieldsSize: MAX_FIELDS_SIZE,
+        maxFilesSize: MAX_FILES_SIZE
+      });
+
+      let [ fields, files ] = await new Promise(resolve => {
+        form.parse(env.origin.req, function (e, fields, files) {
+          if (e) err = e;
+          resolve([ fields, files ]);
+        });
+      });
+
+      if (err) {
+        if (err.statusCode) env.err = err.statusCode;
+        else env.err = { code: N.io.BAD_REQUEST, message: err.message };
+        return;
+      }
+
+      env.req.files  = files;
+      env.req.fields = fields;
+    }
   });
 };
