@@ -22,10 +22,7 @@ module.exports = function (N) {
     if (env.err) return;
 
     // Don't try to render page if body already provided.
-    if (env.body !== null) {
-      if (env.origin.req.method === 'HEAD') env.body = null;
-      return;
-    }
+    if (env.body !== null) return;
 
     var res = env.res,
         view = env.res.view || env.method,
@@ -46,12 +43,6 @@ module.exports = function (N) {
     // Continue with standard http rendering
 
     env.headers['Content-Type'] = 'text/html; charset=UTF-8';
-
-    // If HEAD is requested, it's no need for real rendering.
-    if (env.origin.req.method === 'HEAD') {
-      env.body = null;
-      return;
-    }
 
     // Expose necessary environment/configuration values.
     res.enabled_locales = N.config.locales;
