@@ -3,7 +3,6 @@
 // - env.res
 //   - body
 //   - headers
-//   - log
 
 
 'use strict';
@@ -12,6 +11,7 @@
 const http        = require('http');
 const _           = require('lodash');
 const compression = require('compression')();
+const isFinished  = require('on-finished').isFinished;
 
 
 var valid_headers = {};
@@ -65,18 +65,28 @@ var valid_headers = {};
 
 ////////////////////////////////////////////////////////////////////////////////
 
+// Note, we also listen `responder:bin` to unify logging & error responding
+
+const CHANNELS = [ 'responder:http', 'responder:rpc', 'responder:bin' ];
 
 module.exports = function (N) {
 
-  N.wire.after([ 'responder:http', 'responder:rpc' ], { priority: 100 }, function response_compress(env, callback) {
+  N.wire.after(CHANNELS, { priority: 100 }, function response_compress(env, callback) {
     compression(env.origin.req, env.origin.res, callback);
   });
 
-  N.wire.after([ 'responder:http', 'responder:rpc' ], { priority: 100 }, function response_send(env) {
+  N.wire.after(CHANNELS, { priority: 100 }, function response_send(env) {
     var res = env.origin.res,
         headers = env.headers,
         body = env.body,
         statusCode;
+
+    // If someone already sent reply - do nothing
+    if (isFinished(res)) return;
+
+    // That should not happen, because custom senders should wait
+    // stream end.
+    if (res.headerSent) return;
 
     //
     // Set some obligatory headers
@@ -132,9 +142,6 @@ module.exports = function (N) {
       N.logger.fatal('send_reply: body MUST be a Buffer, String or Null/Undefined. in %s',
                      env.method);
     }
-
-
-    env.log_request(env);
 
     // set status code and send body (if any)
     res.statusCode = env.status;
