@@ -6,162 +6,110 @@
 'use strict';
 
 
-const _      = require('lodash');
-const http   = require('http');
-const encode = require('mdurl').encode;
+const Promise    = require('bluebird');
+const multiparty = require('multiparty');
+const getRawBody = require('raw-body');
 
 
-const MAX_POST_DATA = 10 * 1000 * 1024; // Max post data in bytes
-const MAX_PARAMS_LOG_LENGTH = 60;
-
-// Values of this params should never leak into logs
-const PROTECTED_PARAMS = [ 'pwd', 'pass', 'password' ];
+const MAX_JSON_SIZE  = 100 * 1024;       // 100kb
+const MAX_FILES_SIZE = 10 * 1024 * 1024; // 10mb
 
 
 module.exports = function (N) {
 
-  function formatParams(params) {
-    let result = '';
-
-    _.forEach(params, function (value, key) {
-      result += (result.length === 0) ? '?' : '&';
-      result += encode(String(key), encode.componentChars);
-      result += '=';
-
-      if (PROTECTED_PARAMS.indexOf(key) !== -1) {
-        // Don't dump protected params
-        result += '****';
-
-      } else if (_.isObject(value)) {
-        // Do not dump nested objects.
-        result += '{...}';
-
-      } else {
-        result += encode(String(value), encode.componentChars);
-      }
-
-      if (result.length > MAX_PARAMS_LOG_LENGTH) return false; // terminate
-    });
-
-    return result.slice(0, MAX_PARAMS_LOG_LENGTH);
-  }
-
-
-  function log(env) {
-    var err = env.err,
-        req = env.origin.req,
-        message = http.STATUS_CODES[env.status],
-        logger = N.logger.getLogger('rpc@' + (env.method ? env.method : '')),
-        level = 'info';
-
-    if (err) {
-      if (err.code >= N.io.APP_ERROR) {
-        level = 'fatal';
-        message = err.stack || err.message || JSON.stringify(err);
-      } else if (err.code >= N.io.BAD_REQUEST) {
-        level = 'error';
-        message = err.message || http.STATUS_CODES[err.code];
-      }
-    }
-
-    logger[level]('%s - "RPC %s%s HTTP/%s" - "%s" - %s',
-                  env.req.ip,
-                  env.method || '[empty]',
-                  formatParams(env.params),
-                  req.httpVersion,
-                  req.headers['user-agent'] || '',
-                  message);
-  }
-
-  N.wire.before('responder:rpc', function rpc_prepare(env, callback) {
+  N.wire.before('responder:rpc', async function rpc_prepare(env) {
     let req = env.origin.req;
-
-    env.log_request = log;
-
     //
     // invalid request
     //
 
     if (req.method !== 'POST') {
       env.err = N.io.BAD_REQUEST;
-      callback();
       return;
     }
 
     //
     // Check request size early by header and terminate immediately for big data
     //
-    let length = req.headers['content-length'];
+    let length = parseInt((req.headers['content-length'] || '0'), 10);
 
-    if (!length) {
+    if (!length || isNaN(length)) {
       env.err = N.io.LENGTH_REQUIRED;
-      callback();
       return;
     }
 
-    if (length > MAX_POST_DATA) {
-      env.err = { code: N.io.BAD_REQUEST, message: 'Too big post data' };
-      callback();
+    let err = null;
+    let payload;
+
+    if (/^multipart\/form-data(?:;|$)/i.test(req.headers['content-type'])) {
+      let form = new multiparty.Form({ maxFieldsSize: MAX_JSON_SIZE, maxFilesSize: MAX_FILES_SIZE });
+
+      let [ fields, files ] = await new Promise(resolve => {
+        form.parse(env.origin.req, function (e, fields, files) {
+          if (e) err = e;
+          resolve([ fields, files ]);
+        });
+      });
+
+      if (err) {
+        if (err.statusCode) env.err = err.statusCode;
+        else env.err = { code: N.io.BAD_REQUEST, message: err.message };
+        return;
+      }
+
+      env.req.files = files;
+
+      payload = fields.__payload__[0];
+    } else {
+      let data = await new Promise(resolve => {
+        getRawBody(req, { encoding: true, limit: MAX_JSON_SIZE, length }, (e, data) => {
+          err = e;
+          resolve(data);
+        });
+      });
+
+      if (err) {
+        if (err.statusCode) env.err = err.statusCode;
+        else env.err = { code: N.io.BAD_REQUEST, message: err.message };
+        return;
+      }
+
+      payload = data;
+    }
+
+    try {
+      payload = JSON.parse(payload);
+    } catch (__) {
+      env.err = { code: N.io.BAD_REQUEST, message: 'Cannot parse post data' };
       return;
     }
 
-    //
-    // start harvesting POST data
-    //
-    let chunks = [],
-        chunksLength = 0;
+    env.params = payload.params || {};
 
-    req.on('data', chunk => {
-      chunks.push(chunk);
-      chunksLength += chunk.length;
-
-      if (chunksLength > MAX_POST_DATA) {
-        // max allowed post data reached, drop request.
-        req.removeAllListeners();
-        req.connection.destroy();
-      }
+    // mark files as "uploaded" for validation purposes
+    // (just consistency check, doesn't prevent users from setting empty string)
+    Object.keys(env.req.files).forEach(field => {
+      env.params[field] = '';
     });
 
-    //
-    // when done (on success) process POST data and handle request
-    //
+    // save CSRF token if it was sent
+    req.csrf = payload.csrf;
 
-    req.on('end', () => {
-      let payload;
+    // invalid payload
+    if (!payload.assets_hash || !payload.method) {
+      env.err = N.io.BAD_REQUEST;
+      return;
+    }
 
-      try {
-        payload = JSON.parse(Buffer.concat(chunks).toString());
-        chunks = [];
-      } catch (err) {
-        env.err = { code: N.io.BAD_REQUEST, message: 'Cannot parse post data' };
-        callback();
-        return;
-      }
+    env.method = payload.method;
 
-      env.params = payload.params || {};
-
-      // save CSRF token if it was sent
-      req.csrf = payload.csrf;
-
-      // invalid payload
-      if (!payload.version_hash || !payload.method) {
-        env.err = N.io.BAD_REQUEST;
-        callback();
-        return;
-      }
-
-      env.method = payload.method;
-
-      // invalid client version.
-      // client will check server version by it's own,
-      // so in fact this error is not used by client
-      if (payload.version_hash !== N.version_hash) {
-        env.err = { code: N.io.BAD_REQUEST, message: 'Client version outdated' };
-        callback();
-        return;
-      }
-
-      callback();
-    });
+    // invalid client version.
+    // client will check server version by it's own,
+    // so in fact this error is not used by client
+    if (payload.assets_hash !== N.assets_hash) {
+      env.err = { code: N.io.BAD_REQUEST, message: 'Client version outdated' };
+      return;
+    }
   });
+
 };
