@@ -19,15 +19,9 @@ module.exports = function (N) {
     // Status is always ok - errors are embedded into payload
     env.status = N.io.OK;
 
-    // HEAD requested - no need for real rendering
-    if (env.origin.req.method === 'HEAD') {
-      env.body = null;
-      return;
-    }
-
     // Replace body with serialized payload
     env.body = JSON.stringify({
-      version_hash: N.version_hash,
+      assets_hash: N.assets_hash,
       error: env.err,
       res: env.err ? null : env.res
     });
