@@ -9,8 +9,8 @@
 
 const path        = require('path');
 const send        = require('send');
-const compression = require('compression')();
-
+const compression = require('util').promisify(require('compression')());
+const onFinished  = require('on-finished');
 
 
 module.exports = function (N) {
@@ -30,23 +30,29 @@ module.exports = function (N) {
   });
 
 
-  N.wire.on('server_bin:core.assets', function asset_file_send(env, callback) {
-    compression(env.origin.req, env.origin.res, callback);
+  N.wire.on('server_bin:core.assets', function asset_file_send(env) {
+    return compression(env.origin.req, env.origin.res);
   });
 
-  N.wire.on('server_bin:core.assets', function asset_file_send(env, callback) {
+  N.wire.on('server_bin:core.assets', async function asset_file_send(env) {
     var req = env.origin.req,
         res = env.origin.res;
 
-    if (req.method !== 'GET' && req.method !== 'HEAD') {
-      callback(N.io.BAD_REQUEST);
-      return;
-    }
+    if (req.method !== 'GET' && req.method !== 'HEAD') throw N.io.BAD_REQUEST;
 
-    send(req, env.params.path, { root, index: false, maxage: Infinity })
-      .on('error', err => { callback(err.status); })
-      .on('directory', () => { callback(N.io.BAD_REQUEST); })
-      .on('end', () => { env.log_request(env); })
-      .pipe(res);
+    await new Promise((resolve, reject) => {
+      let ss = send(req, env.params.path, { root, index: false, maxAge: '1y' });
+
+      // Errors with status are not fatal,
+      // rethrow those up as code, not as Error
+      ss.on('error', err => reject(err.status || err));
+
+      onFinished(res, () => {
+        if (res.statusCode) env.status = res.statusCode;
+        resolve();
+      });
+
+      ss.pipe(res);
+    });
   });
 };
