@@ -4,8 +4,10 @@
 'use strict';
 
 
-var path = require('path');
-var send = require('send');
+const path        = require('path');
+const send        = require('send');
+const Promise     = require('bluebird');
+const onFinished  = require('on-finished');
 
 
 module.exports = function (N) {
@@ -17,7 +19,7 @@ module.exports = function (N) {
     // if someone requests '/myfile.txt?xxxx' instead of '/myfile.txt/
     additionalProperties: true,
     properties: {
-      file: {
+      path: {
         type: 'string',
         required: true
       }
@@ -25,19 +27,25 @@ module.exports = function (N) {
   });
 
 
-  N.wire.on('server_bin:core.static', function static_file_send(env, callback) {
+  N.wire.on('server_bin:core.static', async function static_file_send(env) {
     var req = env.origin.req,
         res = env.origin.res;
 
-    if (req.method !== 'GET' && req.method !== 'HEAD') {
-      callback(N.io.BAD_REQUEST);
-      return;
-    }
+    if (req.method !== 'GET' && req.method !== 'HEAD') throw N.io.BAD_REQUEST;
 
-    send(req, env.params.file, { root, index: false })
-      .on('error', err => { callback(err.status); })
-      .on('directory', () => { callback(N.io.BAD_REQUEST); })
-      .on('end', () => { env.log_request(env); })
-      .pipe(res);
+    await new Promise((resolve, reject) => {
+      let ss = send(req, env.params.path, { root, index: false, maxAge: '7d' });
+
+      // Errors with status are not fatal,
+      // rethrow those up as code, not as Error
+      ss.on('error', err => reject(err.status || err));
+
+      onFinished(res, () => {
+        if (res.statusCode) env.status = res.statusCode;
+        resolve();
+      });
+
+      ss.pipe(res);
+    });
   });
 };
