@@ -5,14 +5,14 @@
 'use strict';
 
 
-var _ = require('lodash');
+const _ = require('lodash');
 
 
 module.exports = function (N) {
 
   N.wire.after('server_chain:http:*', { priority: 80 }, function assets_info_inject(env) {
 
-    var key = env.runtime.locale, assetsMap, stylesheetsMap;
+    let key = env.runtime.locale, assetsMap, stylesheetsMap;
 
     if (!N.assets.distribution[key]) {
       // should never happen
