@@ -1,9 +1,8 @@
 'use strict';
 
 
-const Asset = require('./base');
-const _     = require('lodash');
-const co    = require('bluebird-co').co;
+const _       = require('lodash');
+const Asset   = require('./base');
 
 
 function AssetConcat() {
@@ -42,30 +41,28 @@ AssetConcat.prototype.push = function (asset) {
 
 
 AssetConcat.prototype.resolve = function () {
-  let self = this;
+  if (this.__promise__) return this.__promise__;
 
-  if (self.__promise__) return self.__promise__;
+  this.__promise__ = (async () => {
+    await this.__resolveDependencies__();
 
-  self.__promise__ = co(function* () {
-    yield self.__resolveDependencies__();
+    if (await this.__restore_cache__()) return;
 
-    if (yield self.__restore_cache__()) return;
+    let assets = this.__queue__.map(path => this.__bundler__.findAsset(path));
 
-    let assets = self.__queue__.map(path => self.__bundler__.findAsset(path));
-
-    self.source   = '' + assets
+    this.source   = '' + assets
                            .map(asset => asset.source)
                            .filter(src => Boolean(src)) // Exclude empty
                            .join('\n\n');
 
-    yield self.__run_plugins__();
+    await this.__run_plugins__();
 
-    self.digest = self.dependenciesDigest;
+    this.digest = this.dependenciesDigest;
 
-    yield self.__save_cache__();
-  }).then(() => { self.resolved = true; });
+    await this.__save_cache__();
+  })().then(() => { this.resolved = true; });
 
-  return self.__promise__;
+  return this.__promise__;
 };
 
 
