@@ -1,9 +1,8 @@
 'use strict';
 
 
-const Asset = require('./base');
-const path  = require('path');
-const co    = require('bluebird-co').co;
+const path    = require('path');
+const Asset   = require('./base');
 
 
 function AssetFile() {
@@ -25,29 +24,28 @@ Object.defineProperty(AssetFile.prototype, 'digestPath', {
   get() {
     let path_obj = path.parse(this.destPath);
 
-    return path.join('public', path_obj.name + '-' + this.digest + path_obj.ext);
+    // use forward slashes on windows to simplify code
+    return 'public/' + path_obj.name + '-' + this.digest + path_obj.ext;
   }
 });
 
 
 AssetFile.prototype.resolve = function () {
-  let self = this;
+  if (this.__promise__) return this.__promise__;
 
-  if (self.__promise__) return self.__promise__;
+  this.__promise__ = (async () => {
+    await this.__resolveDependencies__();
 
-  self.__promise__ = co(function* () {
-    yield self.__resolveDependencies__();
+    if (await this.__restore_cache__()) return;
 
-    if (yield self.__restore_cache__()) return;
+    await this.__run_plugins__();
 
-    yield self.__run_plugins__();
+    this.digest = this.dependenciesDigest;
 
-    self.digest = self.dependenciesDigest;
+    await this.__save_cache__();
+  })().then(() => { this.resolved = true; });
 
-    yield self.__save_cache__();
-  }).then(() => { self.resolved = true; });
-
-  return self.__promise__;
+  return this.__promise__;
 };
 
 
