@@ -1,9 +1,8 @@
 'use strict';
 
 
-const Asset     = require('./base');
 const BabelFish = require('babelfish');
-const co        = require('bluebird-co').co;
+const Asset     = require('./base');
 
 
 function AssetLang(bundler, options) {
@@ -28,33 +27,31 @@ AssetLang.prototype.addPhrase = function (lang, path, apiPath, phrase) {
 
 
 AssetLang.prototype.resolve = function () {
-  let self = this;
+  if (this.__promise__) return this.__promise__;
 
-  if (self.__promise__) return self.__promise__;
+  this.__promise__ = (async () => {
+    if (await this.__restore_cache__()) return;
 
-  self.__promise__ = co(function* () {
-    if (yield self.__restore_cache__()) return;
-
-    if (self.__queue__.length) {
+    if (this.__queue__.length) {
       let babelfish = new BabelFish();
 
-      if (self.fallback) {
-        babelfish.setFallback(self.lang, self.fallback);
+      if (this.fallback) {
+        babelfish.setFallback(this.lang, this.fallback);
       }
-      self.__queue__.forEach(p => babelfish.addPhrase(p[0], p[1], p[2], false));
-      self.source = babelfish.stringify(self.lang);
+      this.__queue__.forEach(p => babelfish.addPhrase(p[0], p[1], p[2], false));
+      this.source = babelfish.stringify(this.lang);
     } else {
-      self.source = '';
+      this.source = '';
     }
 
-    yield self.__run_plugins__();
+    await this.__run_plugins__();
 
-    self.digest = self.dependenciesDigest;
+    this.digest = this.dependenciesDigest;
 
-    yield self.__save_cache__();
-  }).then(() => { self.resolved = true; });
+    await this.__save_cache__();
+  })().then(() => { this.resolved = true; });
 
-  return self.__promise__;
+  return this.__promise__;
 };
 
 
