@@ -9,7 +9,8 @@ const map = {
   '.less': 'less',
   '.styl': 'stylus',
   '.scss': 'sass',
-  '.jade': 'jade'
+  '.jade': 'jade',
+  '.pug':  'pug'
 };
 
 
