@@ -1,19 +1,12 @@
 'use strict';
 
 
-const autoprefixer = require('autoprefixer');
-const postcss      = require('postcss');
-const Promise      = require('bluebird');
+module.exports = async function (context) {
+  const autoprefixer = require('autoprefixer');
+  const postcss      = require('postcss');
 
+  let ap = postcss([ autoprefixer() ]);
 
-module.exports = function (context) {
-
-  return Promise.resolve().then(() => {
-    // Defaults + Opera 12.1 (still > 1% in Russia)
-    let requirements = '> 1%, last 2 versions, Firefox ESR, Opera 12.1';
-
-    let ap = postcss([ autoprefixer({ browsers: requirements }) ]);
-
-    context.asset.source = ap.process(context.asset.source).css;
-  });
+  /* eslint-disable no-undefined */
+  context.asset.source = (await ap.process(context.asset.source, { from: undefined })).css;
 };
