@@ -1,13 +1,14 @@
 'use strict';
 
 
-const jade    = require('jade');
 const Promise = require('bluebird');
 
 
 module.exports = function (context) {
 
   return Promise.resolve().then(() => {
+    const jade = require('jade');
+
     let tpl = jade.compileClient(context.asset.source, {
       // debug decrease speed and disable external cache
       // (injects unique paths every time, because of random tml dir)
