@@ -1,20 +1,18 @@
 'use strict';
 
 
-const less    = require('less');
 const path    = require('path');
-const Promise = require('bluebird');
 
 
 module.exports = function (context) {
   let node_modules_abs_path = path.join(path.dirname(require.resolve('less')), '../');
 
-  return Promise.fromCallback(cb => {
-    less.render(context.asset.source, {
-      paths: [ path.dirname(context.asset.logicalPath), node_modules_abs_path ],
-      optimization: 1,
-      filename: context.asset.logicalPath
-    }, cb);
+  const less = require('less');
+
+  return less.render(context.asset.source, {
+    paths: [ path.dirname(context.asset.logicalPath), node_modules_abs_path ],
+    optimization: 1,
+    filename: context.asset.logicalPath
   })
   .then(data => {
     context.asset.source = data.css;
