@@ -1,9 +1,7 @@
 'use strict';
 
 
-const sass    = require('node-sass');
-const path    = require('path');
-const Promise = require('bluebird');
+const path = require('path');
 
 
 function resolvePath(file) {
@@ -50,12 +48,12 @@ module.exports = function (context) {
   }
 
 
-  return Promise.fromCallback(cb => {
-    sass.render({
-      data: context.asset.source,
-      file: context.asset.logicalPath,
-      importer
-    }, cb);
+  const sassRender = require('util').promisify(require('node-sass').render);
+
+  return sassRender({
+    data: context.asset.source,
+    file: context.asset.logicalPath,
+    importer
   })
   .then(data => {
     context.asset.source = data.css.toString();
