@@ -1,10 +1,8 @@
 'use strict';
 
 
-const stylus  = require('stylus');
 const path    = require('path');
 const _       = require('lodash');
-const Promise = require('bluebird');
 
 
 function resolvePath(file) {
@@ -22,11 +20,11 @@ function resolvePath(file) {
 }
 
 
-let origFind = stylus.utils.find;
-
-
 module.exports = function (context) {
   return Promise.resolve().then(() => {
+    const stylus   = require('stylus');
+    const origFind = stylus.utils.find;
+
     // monkey-patch lookup with resolver
     stylus.utils.find = function (lookupFile, lookupPaths, thisFilename) {
       return origFind(resolvePath(lookupFile), lookupPaths, thisFilename);
@@ -42,13 +40,10 @@ module.exports = function (context) {
       }
     });
 
-    return Promise.fromCallback(cb => style.render(cb)).then(css => {
-      // add Stylus `@import`s as dependencies of current asset
-      _.forEach(style.options._imports, imported => {
-        context.asset.dependOnFile(imported.path);
-      });
-
-      context.asset.source = css;
+    context.asset.source = style.render();
+    // add Stylus `@import`s as dependencies of current asset
+    _.forEach(style.options._imports, imported => {
+      context.asset.dependOnFile(imported.path);
     });
   });
 };
