@@ -5,9 +5,7 @@ module.exports = function (N) {
   N.runtime = N.runtime || {};
 
   N.wire    = require('event-wire')();
-  N.router  = require('pointer')('$$ N.router.stringify() $$');
-  N.environment  = '$$ JSON.stringify(N.environment) $$';
-  N.version_hash = '$$ JSON.stringify(N.version_hash) $$';
+  N.router  = require('../router')('$$ N.router.stringify() $$');
   // No need to set default locale in constructor, since we use 1
   // language on client.
   N.i18n    = require('babelfish')('en-US');
@@ -21,9 +19,6 @@ module.exports = function (N) {
   N.io             = require('./kernel/io')(N);
   N.runtime.render = require('./kernel/render')(N);
 
-  // refer runtime in templates wrappers. Needed to render templates.
-  N.__jade_runtime = require('jade/lib/runtime.js');
-
   //
   // Emit io.* events like rpc does to signal user about executing
   // network requests.
@@ -31,16 +26,14 @@ module.exports = function (N) {
   let defaultLoadAssets = N.loader.loadAssets;
 
   N.loader.loadAssets = function () {
-    let args = Array.prototype.slice.call(arguments),
-        callback = args[args.length - 1];
-
     N.wire.emit('io.request');
 
-    args[args.length - 1] = function () {
-      N.wire.emit('io.complete', {});
-      return callback.apply(this, arguments);
-    };
-
-    return defaultLoadAssets.apply(this, args);
+    return defaultLoadAssets.apply(this, arguments)
+      .then(function () {
+        N.wire.emit('io.complete', {});
+      }, function (err) {
+        N.wire.emit('io.complete', {});
+        throw err;
+      });
   };
 };
