@@ -27,6 +27,10 @@ module.exports = function (N) {
   //
   // Returned promise extended with `.cancel()` that allow abort XHR request.
   //
+  // Params: an arbitrary JSON object, File or Blob entries are allowed on top
+  // level of said object (serialized into json without files or multipart form
+  // with files).
+  //
   // Options:
   //
   // - persistent - optional, do not terminate request by `navigate.exit`, default `false`
@@ -52,12 +56,59 @@ module.exports = function (N) {
       N.wire.emit('io.request');
     }
 
-    let xhr = $.post('/io/rpc', JSON.stringify({
-      version_hash: N.version_hash,
+    let files = [];
+    let xhr;
+    let contentType;
+    let formData;
+
+    let data = JSON.stringify({
+      assets_hash: N.runtime.assets_hash,
       method: name,
-      csrf: N.runtime.token_csrf,
+      csrf: document.cookie.replace(/(?:(?:^|.*;\s*)csrf-token\s*\=\s*([^;]*).*$)|^.*$/, '$1'),
       params
-    }));
+    }, function find_files(key, value) {
+      // assume that File is inherited from Blob
+      if (value instanceof Blob) {
+        files.push({ key, value });
+
+        /* eslint-disable no-undefined */
+        return undefined;
+      }
+
+      return value;
+    });
+
+    if (files.length) {
+      formData = new FormData();
+      formData.append('__payload__', data);
+      files.forEach(file => formData.append(file.key, file.value));
+      contentType = false; // will be auto-set to "multipart/form-data; boundary=..."
+    } else {
+      formData = data;
+      contentType = 'application/json';
+    }
+
+    xhr = $.ajax({
+      url: '/io/rpc',
+      type: 'POST',
+      data: formData,
+      dataType: 'json',
+      contentType,
+      // turn off replacing encoded spaces like '%20' with '+', see
+      // https://github.com/jquery/jquery/issues/2658
+      processData: false,
+      xhr() {
+        let xhr = $.ajaxSettings.xhr();
+
+        if (options.onProgress) {
+          if (xhr.upload) {
+            xhr.upload.addEventListener('progress', options.onProgress);
+          }
+        }
+
+        return xhr;
+      }
+    });
 
     if (!options.persistent) {
       requests.push(xhr);
@@ -68,9 +119,10 @@ module.exports = function (N) {
 
       data = data || {};
 
-      if (data.version_hash !== N.version_hash) {
+      if (data.assets_hash !== N.runtime.assets_hash) {
         data.error = {
           code: codes.EWRONGVER,
+          hash: data.assets_hash,
           message: 'Server software was updated. Please reload your page to continue.'
         };
         delete data.res;
@@ -78,9 +130,6 @@ module.exports = function (N) {
 
       // If invalid CSRF token error and retry is allowed.
       if (data.error && codes.INVALID_CSRF_TOKEN === data.error.code && options._retryOnCsrfError) {
-        // Renew CSRF token.
-        N.runtime.token_csrf = data.error.data.token;
-
         // Only one attempt to retry is allowed.
         options._retryOnCsrfError = false;
 
