@@ -30,7 +30,7 @@ module.exports = function (N) {
 
     add_raw_data: function add_raw_data_helper(key, data) {
       if (N.runtime.page_data.hasOwnProperty(key)) {
-        N.logger.warn('Override of %j key in N.runtime.page_data');
+        N.logger.warn('Override of %j key in N.runtime.page_data', key);
       }
 
       N.runtime.page_data[key] = data;
