@@ -63,7 +63,7 @@ let DEFAULT_HELPERS = {
 
   add_raw_data: function add_raw_data_helper(key, data) {
     if (this.runtime.page_data.hasOwnProperty(key)) {
-      this.__N.logger.warn('Override of %j key in env.runtime.page_data');
+      this.__N.logger.warn('Override of %j key in env.runtime.page_data', key);
     }
 
     this.runtime.page_data[key] = data;
@@ -105,14 +105,13 @@ function Env(N, options) {
   this.body     = null;
   // internal, http status code (set in responders)
   this.status   = N.io.OK;
-  // filled by formidable with `fields & `files` structures, for http POST requests
-  this.post     = null;
 
   this.headers  = {};
 
 
   this.runtime  = {
-    page_data: {}
+    assets_hash: N.assets_hash,
+    page_data:   {}
   };
 
   let puncher = new Puncher();
@@ -132,7 +131,9 @@ function Env(N, options) {
     isEncrypted: options.isEncrypted,
     ip:          options.remoteAddress,
     matched:     options.matched || null,
-    tzOffset:    new Date().getTimezoneOffset() // use server tz by default
+    tzOffset:    new Date().getTimezoneOffset(), // use server tz by default
+    files:       {}, // uploaded files, fieldName => [ { name, size, path, ... } ]
+    fields:      {}  // form fields if sent via HTTP POST
   };
 
   this.res      = {
@@ -150,8 +151,6 @@ function Env(N, options) {
     },
     menus: {}
   };
-
-  this.log_request = function log_request_dummy() {};
 
   // Pin helpers (those are shared with server methods and renderer)
   //
