@@ -6,6 +6,7 @@ const _             = require('lodash');
 const fs            = require('fs');
 const path          = require('path');
 const cluster       = require('cluster');
+const crypto        = require('crypto');
 const BabelFish     = require('babelfish');
 const Bundler       = require('../../bundler');
 const routerInit    = require('./bundle/router');
@@ -43,7 +44,8 @@ module.exports = function (N) {
         throw `Server module must return an initilizer function as the exports at ${file_path}`;
       }
 
-      let api_path_start = path.dirname(name).replace(new RegExp(path.sep, 'g'), '.');
+      // path is from glob, which always returns forward slashes
+      let api_path_start = path.dirname(name).replace(/\//g, '.');
       let path_obj = path.parse(file_path);
       let api_path;
 
@@ -83,7 +85,7 @@ module.exports = function (N) {
 
   // Process bundle queue
   //
-  N.wire.on('init:bundle', function* bundle_all(N) {
+  N.wire.on('init:bundle', async function bundle_all(N) {
     // In child process just load assets from manifest
     if (cluster.isWorker) return;
 
@@ -104,7 +106,7 @@ module.exports = function (N) {
 
       N.wire.on(`init:bundle.${task.name}`, task.fn);
 
-      yield N.wire.emit(`init:bundle.${task.name}`, sandbox);
+      await N.wire.emit(`init:bundle.${task.name}`, sandbox);
 
       N.logger.info(`◦ ${task.name} ${timer.elapsed}`);
     }
@@ -130,13 +132,18 @@ module.exports = function (N) {
       throw new Error(`Bundle: Can't start process - manifest file '${manifest_path}' not exists or broken.`);
     }
 
-    N.assets = manifest;
     N.i18n = new BabelFish();
+    N.assets = manifest;
+    N.assets_hash = crypto.createHash('md5')
+                          .update(N.version_hash)
+                          .update(JSON.stringify(manifest))
+                          .digest('hex');
 
     let server_assets = _.filter(manifest.files, (__, logical_path) => _.startsWith(logical_path, 'server/'));
 
     _.forEach(server_assets, asset_info => {
       // TODO: remove this check when components will be virtual
+      // assume that digestPath has forward slashes even on windows
       if (_.startsWith(asset_info.digestPath, 'server/package-component-')) {
         return; // continue
       }
@@ -160,8 +167,10 @@ module.exports = function (N) {
     let timer = stopwatch();
 
     N.assets.asset_url = _.memoize(function (name) {
+      // Relative paths not supported
       if (name[0] === '.') {
-        name = path.join(N.mainApp.root, name);
+        N.logger.error(`Failed to find asset ${name}`);
+        return '#';
       }
 
       let asset;
@@ -176,6 +185,7 @@ module.exports = function (N) {
         return '#';
       }
 
+      // assume that digestPath has forward slashes even on windows
       return '/assets/' + asset.digestPath.replace(/^public\//, '');
     });
 
@@ -183,6 +193,7 @@ module.exports = function (N) {
     let files_cache = {};
 
     _.mapValues(N.assets.files, (asset_info, logical_path) => {
+      // assume that digestPath has forward slashes even on windows
       if (!_.startsWith(asset_info.digestPath, 'public/')) {
         return; // continue
       }
@@ -191,8 +202,10 @@ module.exports = function (N) {
     });
 
     N.assets.asset_body = _.memoize(function (name) {
+      // Relative paths not supported
       if (name[0] === '.') {
-        name = path.join(N.mainApp.root, name);
+        N.logger.error(`Failed to find asset ${name}`);
+        return null;
       }
 
       let asset_data = files_cache[require.resolve(name)];
