@@ -4,7 +4,7 @@
 const _ = require('lodash');
 
 
-module.exports = function* (sandbox) {
+module.exports = async function (sandbox) {
   function report_multi_bundled() {
     // module_path -> [ pkg_name ]
     let included_in = {};
@@ -28,7 +28,7 @@ module.exports = function* (sandbox) {
     });
   }
 
-  sandbox.files = yield sandbox.bundler.compile();
+  sandbox.files = await sandbox.bundler.compile();
 
   report_multi_bundled();
 };
