@@ -50,7 +50,7 @@ module.exports = function (sandbox) {
     sandbox.component_server[pkg_name].views = sandbox.bundler.createClass('concat', {
       logicalPath: 'internal/server/package-component-views-' + pkg_name + '.js',
       plugins: [ 'wrapper' ],
-      wrapBefore: 'var jade = require("jade/lib/runtime");\nN.views = N.views || {};\n',
+      wrapBefore: 'N.views = N.views || {};\n',
       wrapAfter: ''
       // TODO: add `virtual: true` when bundler will be finished
     });
@@ -58,7 +58,7 @@ module.exports = function (sandbox) {
     sandbox.component_client[pkg_name].views = sandbox.bundler.createClass('concat', {
       logicalPath: 'internal/public/package-component-views-' + pkg_name + '.js',
       plugins: [ 'wrapper' ],
-      wrapBefore: 'NodecaLoader.execute(function (N) {\nvar jade = N.__jade_runtime;\nN.views = N.views || {};\n',
+      wrapBefore: 'NodecaLoader.execute(function (N, require) {\nN.views = N.views || {};\n',
       wrapAfter: '});'
       // TODO: add `virtual: true` when bundler will be finished
     });
