@@ -25,8 +25,15 @@ module.exports = function (sandbox) {
         package_css.push(sandbox.bundler.createClass('file', {
           logicalPath: file_info.path,
           virtual: true,
-          plugins: [ 'load_text', 'macros', 'auto' ].concat(sandbox.compression ? [ 'csswring' ] : [])
+          plugins: [ 'load_text', 'macros', 'auto' ].concat(sandbox.compression ? [ 'clean-css' ] : [])
         }));
+      });
+
+      _.forEach(sandbox.config.packages[pkg_name].files.css, file_info => {
+        sandbox.bundler.createClass('file', {
+          logicalPath: file_info.path,
+          plugins: [ 'load_text', 'macros', 'auto' ].concat(sandbox.compression ? [ 'clean-css' ] : [])
+        });
       });
     });
   });
