@@ -8,8 +8,9 @@ const uglifyify  = require('uglifyify');
 const stream     = require('readable-stream');
 const path       = require('path');
 const babelify   = require('babelify');
-const utils      = require('./utils/search_assets');
 const Promise    = require('bluebird');
+
+const resolve_module_path = require('./utils/resolve_module_path');
 
 
 const before = _.template(`
@@ -32,54 +33,53 @@ module.exports = function (sandbox) {
   //
   sandbox.included_modules = {};
 
-  function browserify_concat_plugin(context) {
-    return Promise.resolve().then(() => {
-      if (!_.trim(context.asset.source, '\n')) return;
+  async function browserify_concat_plugin(context) {
+    if (!_.trim(context.asset.source, '\n')) return;
 
-      let b = browserify({
-        prelude: 'NodecaLoader.wrap',
-        detectGlobals: false
-      });
-
-      context.asset.__queue__.forEach(file_path => {
-        let s = new stream.Transform();
-        s.push(context.bundler.findAsset(file_path).source);
-        s.end();
-
-        b.add(s, { basedir: path.dirname(file_path), file: file_path });
-      });
-
-      // babelify settings defined in package.json of each application
-      b.transform(cacheify(babelify.configure(), sandbox.cache_db));
-
-      if (sandbox.compression) {
-        b.transform(cacheify(uglifyify, sandbox.cache_db), { global: true });
-      }
-
-      exclude.forEach(e => b.exclude(e));
-
-      b.on('file', path => {
-        context.asset.dependOnFile(path);
-      });
-
-      b.on('dep', function (row) {
-        if (row.entry) {
-          _.forEach(row.indexDeps, (index, dep_name) => {
-            if (index) {
-              let dep_path = utils.resolve_module_path(path.dirname(row.file), dep_name);
-
-              sandbox.included_modules[row.file] = sandbox.included_modules[row.file] || [];
-              sandbox.included_modules[row.file].push(dep_path);
-            }
-          });
-        }
-      });
-
-      return Promise.fromCallback(cb => b.bundle(cb))
-        .then(out => {
-          context.asset.source = String(out);
-        });
+    let b = browserify({
+      prelude: 'NodecaLoader.wrap',
+      detectGlobals: false
     });
+
+    context.asset.__queue__.forEach(file_path => {
+      let s = new stream.Transform();
+      s.push(context.bundler.findAsset(file_path).source);
+      s.end();
+
+      b.add(s, { basedir: path.dirname(file_path), file: file_path });
+    });
+
+    // babelify settings defined in package.json of each application
+    b.transform(cacheify(babelify.configure(), sandbox.cache_db));
+
+    if (sandbox.compression) {
+      b.transform(cacheify(uglifyify, sandbox.cache_db), { global: true });
+    }
+
+    exclude.forEach(e => b.exclude(e));
+
+    b.on('file', path => {
+      context.asset.dependOnFile(path);
+    });
+
+    b.on('dep', function (row) {
+      if (row.entry) {
+        _.forEach(row.indexDeps, (index, dep_name) => {
+          if (index) {
+            let dep_path = resolve_module_path(path.dirname(row.file), dep_name);
+
+            sandbox.included_modules[row.file] = sandbox.included_modules[row.file] || [];
+            sandbox.included_modules[row.file].push(dep_path);
+          }
+        });
+      }
+    });
+
+    let out = await new Promise((resolve, reject) => {
+      b.bundle((err, result) => (err ? reject(err) : resolve(result)));
+    });
+
+    context.asset.source = String(out);
   }
 
   _.forEach(sandbox.config.packages, (pkg, pkg_name) => {
