@@ -2,8 +2,7 @@
 
 
 const path    = require('path');
-const write   = require('write-file-atomic');
-const Promise = require('bluebird');
+const write   = require('util').promisify(require('write-file-atomic'));
 
 
 module.exports = function (sandbox) {
@@ -15,6 +14,5 @@ module.exports = function (sandbox) {
     distribution: sandbox.assets_map
   };
 
-  return Promise.fromCallback(cb =>
-    write(manifest_path, JSON.stringify(manifest, null, 2), cb));
+  return write(manifest_path, JSON.stringify(manifest, null, 2));
 };
