@@ -8,11 +8,11 @@
 // 3rd-party
 const url     = require('url');
 const _       = require('lodash');
-const Pointer = require('pointer');
 
 
 // internal
 const stopwatch = require('../utils/stopwatch');
+const Router    = require('../../router');
 
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -272,8 +272,8 @@ module.exports = function (N) {
   }
 
   // Expose the result, overriding built-in URL parser in pointer with node.js parser
-  N.router = new Pointer(null, u => url.parse(u, false, true));
-  // N.router = new Pointer(null);
+  N.router = new Router(null, u => url.parse(u, false, true));
+  // N.router = new Router(null);
 
   _.forEach(routeConfigs, route => { N.router.addRoute(route); });
 
