@@ -34,29 +34,22 @@ module.exports = function (sandbox) {
   //
   bundler.registerHelper('N', N);
 
-  bundler.registerHelper('asset_url', function asset_url(name) {
-    if (name[0] === '.') {
-      name = path.join(N.mainApp.root, name);
-    }
-
-    let resolved_path = require.resolve(name);
+  bundler.registerHelper('asset_url', function asset_url(asset_path) {
+    let resolved_path = require.resolve(asset_path);
     let asset = bundler.findAsset(resolved_path);
 
     if (!asset) throw new Error(`Invalid asset path ${resolved_path}`);
 
     if (asset.resolved) {
+      // assume that digestPath has forward slashes even on windows
       return '/assets/' + asset.digestPath.replace(/^public\//, '');
     }
 
     throw new bundler.DependencyError(resolved_path);
   });
 
-  bundler.registerHelper('asset_body', function asset_body(name) {
-    if (name[0] === '.') {
-      name = path.join(N.mainApp.root, name);
-    }
-
-    let resolved_path = require.resolve(name);
+  bundler.registerHelper('asset_body', function asset_body(path) {
+    let resolved_path = require.resolve(path);
     let asset = bundler.findAsset(resolved_path);
 
     if (!asset) throw new Error(`Invalid asset path ${resolved_path}`);
