@@ -6,119 +6,13 @@ const fs      = require('fs');
 const path    = require('path');
 const yaml    = require('js-yaml');
 const glob    = require('glob').sync;
-const cached  = require('./fs_cached');
-//  Alternative to node's `require.resolve`, with ability to define root path
-//  + memoise, that boosts speed
-const resolve = _.memoize(require('resolve').sync, JSON.stringify);
-
-
-// Resolve module path
-//
-function resolve_module_path(fromRoot, toFile) {
-  let first = toFile.split(/[\/\\]/)[0];
-
-  if (first === '.' || first === '..') {
-    toFile = path.resolve(fromRoot, toFile);
-  }
-
-  try {
-    return resolve(toFile, { basedir: fromRoot });
-  } catch (err) {
-    return null;
-  }
-}
-
-
-
-// Returns normalized version of vendor section of a package definition.
-// Resolves all file paths relative to `app.root` using `resolveModulePath`.
-//
-// NOTE: Normalized vendor section is different from the config origin.
-//
-// Config:
-//
-// ```
-// vendor:
-//   - lodash
-//   - bar: "./path/to/bar.js"
-// ```
-//
-// Normalized:
-//
-// ```
-// vendor:
-//   "lodash": "/absolute/path/to/lodash.js"
-//   "bar": "/absolute/path/to/bar.js"
-// ```
-//
-function normalize_vendor(app, config) {
-  let result = {};
-
-  _.forEach(config, vendorPkg => {
-    let filename, moduleName, resolvedPath;
-
-    if (_.isPlainObject(vendorPkg)) {
-      moduleName = _.keys(vendorPkg);
-
-      if (moduleName.length === 1) {
-        moduleName = moduleName[0];
-        filename = vendorPkg[moduleName];
-      } else {
-        throw new Error('Ill-formed list of vendor files.');
-      }
-    } else {
-      moduleName = vendorPkg;
-      filename = vendorPkg;
-    }
-
-    resolvedPath = resolve_module_path(app.root, filename);
-
-    if (!resolvedPath) {
-      throw `Bundler cannot find declared vendor file "${filename}"`;
-    }
-
-    result[moduleName] = resolvedPath;
-  });
-
-  return result;
-}
-
-
-function normalize_pkg_config(app, config) {
-  config = _.defaults(config, {
-    vendor: [],
-    entries: [],
-    depends: []
-  });
-
-  config.vendor = _.isArray(config.vendor) ? config.vendor : [ config.vendor ];
-  config.entries = _.isArray(config.entries) ? config.entries : [ config.entries ];
-  config.depends = _.isArray(config.depends) ? config.depends : [ config.depends ];
-
-  config.vendor = normalize_vendor(app, config.vendor);
-
-  config.entries = config.entries.map(entry_path => path.join(app.root, entry_path));
-
-  return config;
-}
 
 
 function load_root_apps_config(applications) {
   let result = {};
 
   applications.forEach(app => {
-    let config;
-    let file_path = path.join(app.root, 'bundle.yml');
-
-    try {
-      config = cached.yaml(file_path);
-    } catch (__) {
-      return; // continue
-    }
-
-    config.packages = _.mapValues(config.packages, pkg => normalize_pkg_config(app, pkg));
-
-    result = _.mergeWith(result, config, (a, b) => {
+    result = _.mergeWith(result, app.config, (a, b) => {
       if (_.isArray(a)) {
         return _.uniq(a.concat(b || []));
       }
@@ -229,7 +123,7 @@ function scan_entry(options) {
   let api_prefix = options.pkg_name + '.' +
     path
       .relative(options.api_path_root, options.entry_root)
-      .replace(new RegExp(path.sep, 'g'), '.');
+      .replace(new RegExp('\\' + path.sep, 'g'), '.');
 
   api_prefix = api_prefix.replace(/[.]$/, ''); // Cut tailing . if nested path empty
 
@@ -318,6 +212,3 @@ module.exports = function (applications) {
 
   return result;
 };
-
-
-module.exports.resolve_module_path = resolve_module_path;
