@@ -2,13 +2,12 @@
 
 
 const _          = require('lodash');
-const Promise    = require('bluebird');
 const browserify = require('browserify');
 const cacheify   = require('cacheify');
 const uglifyify  = require('uglifyify');
 
 
-module.exports = function* (sandbox) {
+module.exports = async function (sandbox) {
   let pkg_names = _.keys(sandbox.config.packages);
 
   for (let i = 0; i < pkg_names.length; i++) {
@@ -19,6 +18,15 @@ module.exports = function* (sandbox) {
     let b = browserify({
       prelude: 'NodecaLoader.wrap',
       detectGlobals: false
+    });
+
+    // exclude vendor files included in other bundles
+    _.forEach(sandbox.config.packages, p => {
+      if (p === pkg) return;
+
+      _.forEach(p.vendor, (__, name) => {
+        b.exclude(name);
+      });
     });
 
     if (sandbox.compression) {
@@ -35,8 +43,11 @@ module.exports = function* (sandbox) {
       virtual: true
     });
 
-    asset.source = yield Promise.fromCallback(cb => b.bundle(cb))
-                            .then(out => String(out));
+    let out = await new Promise((resolve, reject) => {
+      b.bundle((err, result) => (err ? reject(err) : resolve(result)));
+    });
+
+    asset.source = String(out);
 
     sandbox.component_client[pkg_names[i]].js.push(asset);
   }
