@@ -23,18 +23,6 @@ function resolveApiPath(path, apiPathStack) {
 }
 
 
-// Creates new object with specified prototype.
-// Alternative to standard `Object.create` for old browsers.
-//
-function createObject(prototype) {
-  function Constructor() {}
-
-  Constructor.prototype = prototype || {};
-
-  return new Constructor();
-}
-
-
 ////////////////////////////////////////////////////////////////////////////////
 
 
@@ -72,7 +60,7 @@ module.exports = function render(N, apiPath, locals, helpers) {
 
   // Internal render closure.
   // It expects `this` to be `self` object of parent template.
-  execute = function execute(executePath, executeLocals) {
+  execute = function _execute(executePath, executeLocals) {
     let templateFn, templateSelf;
 
     if (N.views[executePath]) {
@@ -83,7 +71,7 @@ module.exports = function render(N, apiPath, locals, helpers) {
 
     // Create new `self` context for current template inheriting all data from
     // parent template's `self`.
-    templateSelf = createObject(this);
+    templateSelf = Object.create(this);
 
     if (executeLocals) {
       _.assign(templateSelf, executeLocals);
