@@ -15,10 +15,9 @@
 'use strict';
 
 
-const co          = require('bluebird-co').co;
+const cluster     = require('cluster');
 const Application = require('./runner/application');
 const stopwatch   = require('./init/utils/stopwatch');
-const cluster     = require('cluster');
 
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -116,22 +115,25 @@ exports.bootstrap = function (options) {
   process.on('SIGQUIT', terminate);
 
 
-  co(function* () {
-    // preload & bootstrap
-    yield require('./runner/initialize')(N);
+  async function execute() {
+    try {
+      // preload & bootstrap
+      await require('./runner/initialize')(N);
 
-    // Load 'init:**' events handlers
-    require('./init')(N);
+      // Load 'init:**' events handlers
+      require('./init')(N);
 
-    // execute cli script
-    yield require('./runner/exec_cli')(N);
+      // execute cli script
+      await require('./runner/exec_cli')(N);
 
-    if (cluster.isWorker) {
-      process.send('worker.running');
+      if (cluster.isWorker) {
+        process.send('worker.running');
+      }
+    } catch (err) {
+      logFatalError(err);
+      return N.wire.emit('exit.terminate', 1);
     }
-  })
-  .catch(err => {
-    logFatalError(err);
-    return N.wire.emit('exit.terminate', 1);
-  });
+  }
+
+  execute();
 };
