@@ -6,18 +6,16 @@
 'use strict';
 
 
-const path = require('path');
+const load_app_config = require('../init/bundle/utils/load_app_config');
 
 
 ////////////////////////////////////////////////////////////////////////////////
 
 
 function Application(options) {
-  let pkg = require(path.join(options.root, 'package.json'));
-
   this.root     = options.root;
-  this.name     = pkg.name;
-  this.version  = pkg.version;
+  this.name     = options.name;
+  this.config   = load_app_config(options.root);
 
   this.init     = options.init || function () {};
 }
