@@ -1,7 +1,7 @@
 // Final stage of runner's bootstrap. Searches for all CLI command scripts
 // (`cli/*.js` files) in all applications and then executes requested command:
 //
-//    ./nodeca.js server
+//    ./server.js server
 //
 // will excutes `cli/server.js` command (if any).
 //
