@@ -8,13 +8,12 @@ const cluster     = require('cluster');
 const crypto      = require('crypto');
 const fs          = require('fs');
 const path        = require('path');
-const co          = require('bluebird-co').co;
 const Promise     = require('bluebird');
-const npm         = require('npm');
 
 const _           = require('lodash');
 const log4js      = require('log4js');
 const validator   = require('is-my-json-valid');
+const email_regex = require('email-regex');
 const yaml        = require('js-yaml');
 const wire        = require('event-wire');
 const glob        = require('glob').sync;
@@ -147,8 +146,9 @@ function getAppName(app) {
 // Check list of applications, if application is
 // missed - try to install it via npm.
 //
-const installAppsIfMissed = co.wrap(function* (apps) {
-  let appName, appNpmPath;
+async function installAppsIfMissed(N, apps) {
+  let appName, appNpmPath,
+      missed = [];
 
   for (let app of apps) {
     if (_.isPlainObject(app)) {
@@ -161,12 +161,24 @@ const installAppsIfMissed = co.wrap(function* (apps) {
     try {
       require.resolve(appName);
     } catch (__) {
-      yield Promise.fromCallback(cb => npm.load({}, cb));
-      /*eslint-disable no-loop-func*/
-      yield Promise.fromCallback(cb => npm.commands.install([ appNpmPath ], cb));
+      missed.push(appNpmPath);
     }
   }
-});
+
+  if (missed.length) {
+    /* eslint-disable no-console */
+    console.log(`Missed apps: ${missed.join(', ')}. Installing...`);
+
+    require('child_process').execSync(
+      `yarn add ${missed.join(' ')} --production --non-interactive`,
+      { stdio: 'inherit', cwd: N.mainApp.root }
+    );
+  }
+
+  // in node 6 `require()` for new packages start work in nextTick only.
+  let nextTick = require('util').promisify(process.nextTick);
+  await nextTick();
+}
 
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -181,8 +193,8 @@ const installAppsIfMissed = co.wrap(function* (apps) {
 //
 function initWire(N) {
   N.wire  = wire({
-    p: Promise,
-    co
+    p:  Promise,
+    co: (fn, params) => Promise.coroutine(fn)(params)
   });
 
   function findPuncher(params) {
@@ -216,7 +228,8 @@ function initScope(N) {
 
   // Additional format extentions
   let validateFormatExt = {
-    mongo: /^[0-9a-f]{24}$/
+    mongo: /^[0-9a-f]{24}$/,
+    email: email_regex({ exact: true })
   };
 
   /**
@@ -278,33 +291,9 @@ function initScope(N) {
 }
 
 
-function initConfig(N, mainConfig) {
-  //
-  // Create empty object that we'll fill in a second
-  //
-
-  N.config = {};
-
-  //
-  // Start reading configs:
-  //
-  // - Main app config stored into mainConfig
-  // - Sub-apps configs got merged into N.config
-  // - After all mainConfig got merged into N.config
-  //
-
-  // read configs of sub-applications
-  if (mainConfig.applications && mainConfig.applications.length) {
-    _.forEach(mainConfig.applications, app => {
-      let root = path.join(path.dirname(require.resolve(getAppName(app))), '/config');
-
-      mergeConfigs(N.config, loadConfigs(root));
-    });
-  }
-
-  // merge in main config and resolve `per-environment` configs
-  mergeConfigs(N.config, mainConfig);
-
+// Finalize config, expand environment-dependent configs
+//
+function initConfig(N) {
   // set application environment
   N.environment = process.env.NODECA_ENV || N.config.env_default || 'development';
 
@@ -338,11 +327,9 @@ function initLogger(N) {
       config       = _.assign({}, N.config.logger),
       options      = _.assign({ file: { logSize: 10, backups: 5 } }, config.options),
       // common logging level (minimal threshold)
-      baseLevel    = log4js.levels.toLevel(options.level, log4js.levels.ALL),
+      baseLevel    = log4js.levels.getLevel(options.level, log4js.levels.ALL),
       // should it log everything to console or not
-      logToConsole = (N.environment !== 'production'),
-      // cache of initialized appenders
-      appenders    = {},
+      logToConsole = (N.environment !== 'production' || process.stdout && process.stdout.isTTY),
       // real loggers created for each entry in the config
       loggers      = [],
       // cache of met channels, maps full channel names to corresponding loggers
@@ -357,17 +344,19 @@ function initLogger(N) {
   //  %c - category
   //  %m - message
   //
-  let plainLayout    = log4js.layouts.layout('pattern',
-                         { pattern: '[%d] [%p] %z %c - %m' }
-                       );
+  let plainLayout = {
+    type: 'pattern',
+    pattern: '[%d] [%p] %z %c - %m'
+  };
 
   // Layout for console loggers
   //
   // only difference is `%[`..`%]` - defines highlighted (coloured) part
   //
-  let colouredLayout = log4js.layouts.layout('pattern',
-                         { pattern: '%[[%d] [%p] %z %c -%] %m' }
-                       );
+  let colouredLayout = {
+    type: 'pattern',
+    pattern: '%[[%d] [%p] %z %c -%] %m'
+  };
 
   //
   // define system (general) logger
@@ -381,7 +370,8 @@ function initLogger(N) {
 
   N.logger.setLevel = function (level) {
     level = log4js.levels[level.toUpperCase()];
-    log4js.setGlobalLogLevel(level);
+    N.logger.level = level;
+    for (let logger of loggers) logger.level = level;
   };
 
 
@@ -453,41 +443,14 @@ function initLogger(N) {
   };
 
   //
-  // Load supported appenders
+  // Configure log4js
   //
-
-  log4js.loadAppender('file');
-  log4js.loadAppender('console');
-  log4js.loadAppender('clustered');
-  log4js.loadAppender('logLevelFilter');
-
-  //
-  // clear default loggers
-  //
-
-  log4js.clearAppenders();
-
-  function clusteredAppender(children, category) {
-    let clusterAppenderConfig = {
-      appenders: [],      // configs
-      actualAppenders: [] // functions
-    };
-
-    children.forEach(appender => {
-      clusterAppenderConfig.appenders.push({ category });
-      clusterAppenderConfig.actualAppenders.push(appender);
-    });
-
-    return log4js.appenders.clustered(clusterAppenderConfig);
-  }
-
-  //
-  // configure console logger for non-production environment only
-  //
-
-  if (logToConsole) {
-    log4js.addAppender(log4js.appenders.console(colouredLayout));
-  }
+  let log4js_options = {
+    appenders: {
+      console: { type: 'console', layout: colouredLayout }
+    },
+    categories: {}
+  };
 
   //
   // leave only loggers (with appenders) configs, removing keywords
@@ -498,39 +461,48 @@ function initLogger(N) {
   //
   // configure logger categories and appenders
   //
+  let thresholdAppenderId = 0;
 
   _.forEach(config, (loggerConfig, name) => {
     let appendersInGroup = [];
     let groupLevel = log4js.levels.FATAL;
 
     _.forEach(loggerConfig, appenderConfig => {
-      if (!appenders[appenderConfig.file] && cluster.isMaster) {
+      // add custom prefix "file:" to avoid collisions with console appenders and such
+      if (!log4js_options.appenders['file:' + appenderConfig.file] && cluster.isMaster) {
         let filename = path.resolve(mainRoot, appenderConfig.file);
 
         // make sure destination directory for log file exists
         mkdirp(path.dirname(filename));
 
-        appenders[appenderConfig.file] = log4js.appenders.file(
-          filename,                           // filename
-          plainLayout,                        // layout
-          options.file.logSize * 1024 * 1024, // logSize
-          options.file.backups                // numBackups
-        );
+        log4js_options.appenders['file:' + appenderConfig.file] = {
+          type:       'file',
+          filename,
+          layout:     plainLayout,
+          maxLogSize: options.file.logSize * 1024 * 1024,
+          backups:    options.file.backups
+        };
       }
 
       let myLevel = baseLevel;
 
       if (appenderConfig.level) {
-        myLevel = log4js.levels.toLevel(appenderConfig.level, baseLevel);
+        myLevel = log4js.levels.getLevel(appenderConfig.level, baseLevel);
 
         // get upper threshold
         myLevel = myLevel.isGreaterThanOrEqualTo(baseLevel) ? myLevel : baseLevel;
       }
 
       // return thresholded appender
-      let appender = log4js.appenders.logLevelFilter(myLevel, log4js.levels.FATAL, appenders[appenderConfig.file]);
+      let key = 'threshold:' + thresholdAppenderId++;
 
-      appendersInGroup.push(appender);
+      log4js_options.appenders[key] = {
+        type: 'logLevelFilter',
+        appender: 'file:' + appenderConfig.file,
+        level: myLevel.levelStr
+      };
+
+      appendersInGroup.push(key);
 
       groupLevel = groupLevel.isGreaterThanOrEqualTo(myLevel) ? myLevel : groupLevel;
     });
@@ -544,12 +516,29 @@ function initLogger(N) {
     }
 
     if (appendersInGroup.length) {
-      let clustered = clusteredAppender(appendersInGroup, name);
-      let loglevel  = log4js.appenders.logLevelFilter(groupLevel, log4js.levels.FATAL, clustered);
+      appendersInGroup.forEach(appender => {
+        if (!log4js_options.categories[name]) {
+          log4js_options.categories[name] = { appenders: [], level: 'ALL' };
 
-      log4js.addAppender(loglevel, name);
+          if (logToConsole) {
+            log4js_options.categories[name].appenders.push('console');
+          }
+        }
+
+        log4js_options.categories[name].appenders.push(appender);
+      });
     }
   });
+
+  // alias "default" to "system", log4js requires category named "default"
+  // to be present
+  if (!log4js_options.categories.default) {
+    log4js_options.categories.default = _.cloneDeep(
+      log4js_options.categories.system
+    );
+  }
+
+  log4js.configure(log4js_options);
 
   //
   // Ensure loggers are placed in order from most specific to most general.
@@ -597,26 +586,35 @@ function checkConfig(N) {
 }
 
 
-// Run `init()` method for all registered apps.
-// Usually, hooks loading is placed there
+// Load app and all nested apps, merge their configuration
 //
-function initApps(N) {
+function loadApp(N, appPath) {
+  let app = new Application(require(appPath));
+  let appConfigPath = path.join(app.root, 'config');
+
+  N.apps.push(app);
+  mergeConfigs(N.config, loadConfigs(appConfigPath));
+
+  for (let m of app.config.modules) {
+    loadApp(N, path.join(app.root, m));
+  }
+}
+
+
+// Load configuration for all enabled apps.
+//
+function loadApps(N, applications) {
   N.apps = [ N.mainApp ];
 
   // Try load each enabled application and push to the array of loaded apps
-  _.forEach(N.config.applications, app => {
-    N.apps.push(new Application(require(getAppName(app))));
-  });
-
-  // Call init on each application
-  _.forEach(N.apps, app => app.init(N));
+  _.forEach(applications, appDef => loadApp(N, getAppName(appDef)));
 }
 
 
 ////////////////////////////////////////////////////////////////////////////////
 
 
-module.exports = co.wrap(function* (N) {
+module.exports = async function (N) {
   initScope(N);
   initWire(N);
 
@@ -625,17 +623,30 @@ module.exports = co.wrap(function* (N) {
 
   if (cluster.isMaster) {
     // Ensure all required apps installed before read configs
-    yield installAppsIfMissed(mainConfig.applications || []);
+    await installAppsIfMissed(N, mainConfig.applications || []);
   }
 
-  initConfig(N, mainConfig);
+  // Create empty object that will be filled by each app later
+  N.config = {};
+
+  loadApps(N, mainConfig.applications);
+
+  // merge in main config
+  mergeConfigs(N.config, mainConfig);
+
+  // resolve per-environment configs
+  initConfig(N);
+
   initLogger(N);
 
   N.logger.info('Loaded config files', N.__startupTimer.elapsed);
   let timer = stopwatch();
 
   checkConfig(N);
-  initApps(N);
+
+  // Run `init()` method for all registered apps.
+  // Usually, hooks loading is placed there
+  _.forEach(N.apps, app => app.init(N));
 
   //
   // Create `N.version_hash` - unique value, that tracks packages
@@ -653,7 +664,10 @@ module.exports = co.wrap(function* (N) {
   hasher.update(JSON.stringify(_.omit(N.config, [ 'logger' ])));
 
   N.apps.forEach(app => {
-    hasher.update(fs.readFileSync(path.join(app.root, 'package.json'), 'utf-8'));
+    // `package.json` is not mandatory (likely won't exist for submodules)
+    try {
+      hasher.update(fs.readFileSync(path.join(app.root, 'package.json'), 'utf-8'));
+    } catch (__) {}
 
     // `bundle.yml` is not mandatory
     try {
@@ -664,4 +678,4 @@ module.exports = co.wrap(function* (N) {
   N.version_hash = hasher.digest('hex');
 
   N.logger.info('Applications intialized', timer.elapsed);
-});
+};
