@@ -5,15 +5,9 @@ const _           = require('lodash');
 const fs          = require('mz/fs');
 const crypto      = require('crypto');
 const validator   = require('is-my-json-valid');
-const formidable  = require('formidable');
-const Promise     = require('bluebird');
 
 
 const config_schema = require('./font/_lib/config_schema');
-
-
-const MAX_POST_DATA = 10 * 1000 * 1024; // Max post data in bytes
-const MAX_POST_FIELDS = 20;
 
 
 module.exports = function (N, apiPath) {
@@ -21,103 +15,54 @@ module.exports = function (N, apiPath) {
   N.validate(apiPath, {});
 
 
-  function remove_uploaded_files(files) {
-    _.forEach(files, fileInfo => fs.unlink(fileInfo.path));
-  }
-
-
   // Validate and parse post data
   //
-  N.wire.before(apiPath, function validate_and_parse(env, callback) {
-    // check length
-    let len = parseInt(env.origin.req.headers['content-length'], 10);
+  N.wire.before(apiPath, function validate_and_parse(env) {
+    // Validate post `url` param
+    let fields_validator = validator({
+      properties: {
+        url: {
+          format : 'url',
+          maxLength : 200,
+          required  : false,
+          messages: {
+            format: 'Invalid URL format',
+            maxLength: 'URL too long'
+          }
+        }
+      }
+    }, { verbose: true });
 
-    if (!len || len > MAX_POST_DATA) {
-      callback(413); // Request Entity Too Large
-      return;
+    let url = _.get(env.req, 'fields.url.0');
+
+    if (url && !fields_validator({ url })) {
+      throw { code: N.io.BAD_REQUEST, message: fields_validator.errors[0].message };
     }
 
-    // create & configure form parser
-    let form = new formidable.IncomingForm();
+    // Validate post `config` param
+    let path = _.get(env.req, 'files.config.0.path');
 
-    form.maxFields = MAX_POST_FIELDS;
-
-    // parse form & process result
-    form.parse(env.origin.req, function (err, fields, files) {
-      if (err) {
-        callback(err);
-        remove_uploaded_files(files);
-        return;
-      }
-
-      // Validate post `url` param
-      let fields_validator = validator({
-        properties: {
-          url: {
-            format : 'url',
-            maxLength : 200,
-            required  : false,
-            messages: {
-              format: 'Invalid URL format',
-              maxLength: 'URL too long'
-            }
-          }
-        }
-      }, { verbose: true });
-
-      if (!fields_validator(fields)) {
-        callback({ code: N.io.BAD_REQUEST, message: fields_validator.errors[0].message });
-        remove_uploaded_files(files);
-        return;
-      }
-
-      // Validate post `config` param
-      let files_validator = validator({
-        properties: {
-          config: {
-            type: 'object',
-            required: true,
-            messages: {
-              required: 'Missed "config" param - must be file'
-            }
-          }
-        }
-      }, { verbose: true });
-
-      if (!files_validator(files)) {
-        callback({ code: N.io.BAD_REQUEST, message: files_validator.errors[0].message });
-        remove_uploaded_files(files);
-        return;
-      }
-
-      env.data.files = files;
-      env.data.fields = fields;
-
-      callback();
-    });
+    if (!path) {
+      throw { code: N.io.BAD_REQUEST, message: 'Missed "config" param - must be file' };
+    }
   });
 
 
   // Create session and save config
   //
-  N.wire.on(apiPath, function* save_config(env) {
-    let configPath = env.data.files.config.path;
+  N.wire.on(apiPath, async function save_config(env) {
+    let configPath = _.get(env.req, 'files.config.0.path');
     let configFile;
 
     // Extract config
-    try {
-      configFile = yield fs.readFile(configPath, { encoding: 'utf-8' });
-    } catch (err) {
-      remove_uploaded_files(env.data.files);
-      throw err;
-    }
+    configFile = await fs.readFile(configPath, { encoding: 'utf-8' });
 
     let config;
 
     try {
       config = JSON.parse(configFile);
     } catch (__) {
-      throw { code: N.io.BAD_REQUEST, message: 'Can\'t parse config data' };
+      throw { code: N.io.BAD_REQUEST, message: "Can't parse config data" };
     }
 
     // Validate config content
@@ -141,14 +86,15 @@ module.exports = function (N, apiPath) {
       config
     };
 
-    if (env.data.fields.url) linkData.url = env.data.fields.url;
+    let url = _.get(env.req, 'fields.url.0');
 
-    yield Promise.fromCallback(cb => N.shortlinks.put(
+    if (url) linkData.url = url;
+
+    await N.shortlinks.put(
       sid,
       linkData,
-      { ttl: 6 * 60 * 60 * 1000, valueEncoding: 'json' },
-      cb
-    ));
+      { ttl: 6 * 60 * 60 * 1000, valueEncoding: 'json' }
+    );
 
     throw { code: N.io.OK, message: sid };
   });
