@@ -6,17 +6,19 @@ var request = require('supertest')(TEST.N.config.bind.default.mount);
 
 describe('HTTP alive', function () {
 
-  it('/ GET', function (done) {
-    request
+  it('/ GET', function () {
+    return request
       .get('/')
       .expect(/<!DOCTYPE html>/)
-      .expect(200, done);
+      .expect(200);
   });
 
-  it('/ HEAD', function (done) {
-    request
+  it('/ HEAD', function () {
+    return request
       .head('/')
-      .expect({})
-      .expect(200, done);
+      .expect(200, {})
+      .expect(res => {
+        if (res.text) throw new Error(`Body should be empty, got ${res.text}`);
+      });
   });
 });
